requirements.txt
openai>=1.83.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
python-magic>=0.4.27  # for file type detection 